Move variant dimension values computation to build variants
