+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/internal/BuildVariants.java
@@ -15,13 +15,19 @@
  */
 package dev.nokee.platform.base.internal;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Iterables;
 import dev.nokee.model.internal.core.DescendantNodes;
+import dev.nokee.model.internal.core.GradlePropertyComponent;
 import dev.nokee.model.internal.core.ModelNode;
+import dev.nokee.model.internal.core.ModelPath;
 import dev.nokee.platform.base.BuildVariant;
 import dev.nokee.runtime.core.CoordinateSet;
 import dev.nokee.runtime.core.CoordinateSpace;
+import dev.nokee.utils.TransformerUtils;
 import lombok.val;
+import lombok.var;
 import org.gradle.api.Transformer;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.HasMultipleValues;
@@ -31,10 +37,12 @@ import org.gradle.api.provider.SetProperty;
 
 import java.util.List;
 import java.util.Set;
+import java.util.function.Consumer;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import static dev.nokee.model.internal.core.ModelComponentType.componentOf;
+import static dev.nokee.runtime.core.Coordinates.absentCoordinate;
 import static dev.nokee.utils.Cast.uncheckedCastBecauseOfTypeErasure;
 import static dev.nokee.utils.TransformerUtils.transformEach;
 
@@ -50,7 +58,7 @@ public final class BuildVariants {
 			return dimensionNodes.collect(Collectors.toList());
 		});
 		this.dimensions = dimensions
-			.map(transformEach(it -> it.getComponent(componentOf(VariantDimensionValuesComponent.class)).asProvider()))
+			.map(transformEach(new ToCoordinateSet(entity.getComponent(ModelPath.class).getName())))
 			.flatMap(new ToProviderOfIterableTransformer<>(() -> uncheckedCastBecauseOfTypeErasure(objects.listProperty(CoordinateSet.class))));
 		this.finalSpace = this.dimensions.map(CoordinateSpace::cartesianProduct);
 		this.buildVariants = objects.setProperty(BuildVariant.class);
@@ -73,6 +81,84 @@ public final class BuildVariants {
 
 	public Provider<Set<BuildVariant>> get() {
 		return buildVariants;
+	}
+
+	private static final class ToCoordinateSet implements Transformer<Provider<CoordinateSet<?>>, ModelNode> {
+		private final String componentName;
+
+		public ToCoordinateSet(String componentName) {
+			this.componentName = componentName;
+		}
+
+		@Override
+		public Provider<CoordinateSet<?>> transform(ModelNode entity) {
+			@SuppressWarnings("unchecked")
+			val values = (Provider<Set<Object>>) entity.getComponent(GradlePropertyComponent.class).get();
+
+			return values.map(asCoordinateSet(entity, componentName));
+		}
+
+		private Transformer<CoordinateSet<Object>, Iterable<Object>> asCoordinateSet(ModelNode entity, String componentName) {
+			val axis = entity.getComponent(VariantDimensionAxisComponent.class).get();
+
+			var axisValues = assertNonEmpty(axis.getDisplayName(), componentName);
+
+			var axisCoordinates = axisValues.andThen(transformEach(axis::create));
+
+			val axisValidator = entity.findComponent(VariantDimensionAxisValidatorComponent.class)
+				.map(VariantDimensionAxisValidatorComponent::get);
+			if (axisValidator.isPresent()) {
+				axisCoordinates = axisCoordinates.andThen(new PeekTransformer<>(it -> axisValidator.get().accept(it)));
+			}
+
+			if (entity.hasComponent(VariantDimensionAxisOptionalTag.class)) {
+				axisCoordinates = axisCoordinates.andThen(appended(absentCoordinate(axis)));
+			}
+
+			return axisCoordinates.andThen(CoordinateSet::of);
+		}
+
+		public static <T> TransformerUtils.Transformer<Iterable<T>, Iterable<T>> appended(T element) {
+			return new IterableAppendedAllTransformer<>(ImmutableList.of(element));
+		}
+
+		public static <T> TransformerUtils.Transformer<Iterable<T>, Iterable<T>> appendedAll(Iterable<T> suffix) {
+			return new IterableAppendedAllTransformer<>(suffix);
+		}
+
+		public static final class IterableAppendedAllTransformer<T> implements TransformerUtils.Transformer<Iterable<T>, Iterable<T>> {
+			private final Iterable<T> suffixElements;
+
+			public IterableAppendedAllTransformer(Iterable<T> suffixElements) {
+				this.suffixElements = suffixElements;
+			}
+
+			@Override
+			public Iterable<T> transform(Iterable<T> values) {
+				return Iterables.concat(values, suffixElements);
+			}
+		}
+
+		public static <I extends Iterable<T>, T> TransformerUtils.Transformer<I, I> assertNonEmpty(String propertyName, String componentName) {
+			return new PeekTransformer<>(new AssertNonEmpty<>(propertyName, componentName));
+		}
+
+		private static final class AssertNonEmpty<T> implements Consumer<Iterable<T>> {
+			private final String propertyName;
+			private final String componentName;
+
+			private AssertNonEmpty(String propertyName, String componentName) {
+				this.propertyName = propertyName;
+				this.componentName = componentName;
+			}
+
+			@Override
+			public void accept(Iterable<T> values) {
+				if (Iterables.isEmpty(values)) {
+					throw new IllegalArgumentException(String.format("A %s needs to be specified for component '%s'.", propertyName, componentName));
+				}
+			}
+		}
 	}
 
 	private static final class ToProviderOfIterableTransformer<T, C extends Provider<? extends Iterable<T>> & HasMultipleValues<T>> implements Transformer<Provider<? extends Iterable<T>>, Iterable<Provider<T>>> {
